refactor: move all redux folders to 1 folder redux
