+++ src/HW/client/src/redux/selectors/itemsViewSelectors.js
