+++ src/HW/client/src/store.js
@@ -1,7 +1,7 @@
 
 import { configureStore } from "@reduxjs/toolkit";
 import thunkMiddleware from "redux-thunk";
-import allReducers from "./reducers";
+import allReducers from "./redux/reducers";
 
 export const store = configureStore({
   reducer: allReducers,
